Add files via upload
The complete Final Project that I made for CS50 and this takes me two months
I'm open to crtiticism and suggestion, since I am a beginner, so please tell me if you find anything
Also, I do know that it could/should be better. Some codes are repeated with slight/no changes. Should have just made a function that deals with it.
This was my first project though, so I understand if this was VERY bad to your standard. I hope you'll give me suggestions to make me a better programmer in the future